networking_libraries/strawberry-graphql/pyproject.toml
[tool.poetry]
name = "graphql-queries"
version = "1.0.0"
description = "Examples of GraphQL Querying with Python and the ariadne Package (https://ariadnegraphql.org/)"
authors = [
    "Tommy P. Keane <talk.to.tommypkeane@gmail.com>",
]
maintainers = [
    "Tommy P. Keane <talk.to.tommypkeane@gmail.com>",
]
homepage = "https://www.tommypkeane.com"
repository = "https://github.com/TommyPKeane/learning-python"
license = "MIT"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = "^0.168.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
hypothesis = "^6.70.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
no_lines_before = "LOCALFOLDER"
lines_after_imports = 2

[tool.pytest.ini_options]
minversion = "7.0"
console_output_style = "percent"
log_auto_indent = true
log_level = "DEBUG"
log_cli = true  # Set true if you wanna see the logs live, otherwise look at `log_file`
log_cli_level = "DEBUG"
log_cli_date_format = "%Y-%m-%dT%H:%M:%S%z (%Z)"
log_cli_format = "%(asctime)s :: %(name)s :: %(levelname)s :: %(message)s"
addopts = "--cov"
testpaths = [
    "tests_unit",
]
